requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
whitenoise>=6.7,<7.0
dj-database-url>=2.3,<3.0
django-cors-headers>=4.6,<5.0
gunicorn>=21.2,<22.0
